Issue: WSWRPR-1344 merge from release_rt mass_ market WSWRPR-1314 merge from release_firecrest_r39.p9
Merge branch 'release_rt_mass_market'
Merge branch 'release_firecrest_r39.p9'

+++ wlcmgr/wlan.c
@@ -1619,7 +1619,7 @@ static int do_stop(struct wlan_network *network)
 
     if (network->role == WLAN_BSS_ROLE_UAP)
     {
-        ret = wifi_uap_stop((int)network->type);
+        ret = wifi_uap_stop(network->type);
 
         if (ret != 0)
         {
@@ -4402,7 +4402,7 @@ int wlan_stop(void)
     }
     if (wlan.uap_state > CM_UAP_CONFIGURED)
     {
-        (void)wifi_uap_stop((int)wlan.networks[wlan.cur_uap_network_idx].type);
+        (void)wifi_uap_stop(wlan.networks[wlan.cur_uap_network_idx].type);
         (void)dhcp_server_stop();
     }
